Adding cloud to the weather fetcher
Works like the IP weather fetcher, returning a forecast for an IP address.
This version uses the Temporal Cloud instead of a development server on
your local computer. It is intended to be deployed as an addressable
Web service with a known host/address. It is currently hardwired for
my droplet.